requirements.txt
APScheduler==3.11.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
schedule==1.2.2
tzlocal==5.3.1
urllib3==2.4.0
